Update: Format README to include another dependency and moved login.tsk to app to resolve an error
