--- firebase/login.tsx
@@ -1,38 +1,0 @@
-// app/login.tsx
-import React, { useEffect } from "react";
-import { Button, View, Text } from "react-native";
-import * as WebBrowser from "expo-web-browser";
-import * as Google from "expo-auth-session/providers/google";
-import { GoogleAuthProvider, signInWithCredential } from "firebase/auth";
-import { auth } from "./firebaseConfig";
-
-WebBrowser.maybeCompleteAuthSession();
-
-export default function LoginScreen() {
-  const [request, response, promptAsync] = Google.useIdTokenAuthRequest({
-    clientId: "YOUR_GOOGLE_CLIENT_ID.apps.googleusercontent.com", // from Google Cloud console
-  });
-
-  useEffect(() => {
-    if (response?.type === "success") {
-      const { id_token } = response.params;
-      const credential = GoogleAuthProvider.credential(id_token);
-      signInWithCredential(auth, credential)
-        .then((result) => {
-          console.log("User signed in:", result.user);
-        })
-        .catch((err) => console.error("Auth error:", err));
-    }
-  }, [response]);
-
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>Login with Google</Text>
-      <Button
-        disabled={!request}
-        title="Sign in with Google"
-        onPress={() => promptAsync()}
-      />
-    </View>
-  );
-}
